Refactor: 명칭 통일
